Log and serialization definitions in a new module
